ST-492 Refactoring integration tests from divorce to sptribs (#22)
