+++ src/integrationTest/java/uk/gov/hmcts/sptribs/document/content/GeneralOrderTemplateContentIT.java
@@ -1,23 +1,23 @@
-package uk.gov.hmcts.divorce.document.content;
+package uk.gov.hmcts.sptribs.document.content;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.divorce.ciccase.model.CaseData;
+import uk.gov.hmcts.sptribs.ciccase.model.CaseData;
 
 import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.entry;
-import static uk.gov.hmcts.divorce.ciccase.model.ApplicationType.JOINT_APPLICATION;
-import static uk.gov.hmcts.divorce.ciccase.model.GeneralOrderJudgeOrLegalAdvisorType.ASSISTANT_JUSTICES_CLERK;
-import static uk.gov.hmcts.divorce.ciccase.model.GeneralOrderJudgeOrLegalAdvisorType.PROPER_OFFICER_OF_THE_COURT;
-import static uk.gov.hmcts.divorce.document.content.DocmosisTemplateConstants.GENERAL_ORDER_MADE_BY;
-import static uk.gov.hmcts.divorce.testutil.TestConstants.TEST_CASE_ID;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.caseData;
-import static uk.gov.hmcts.divorce.testutil.TestDataHelper.getGeneralOrder;
+import static uk.gov.hmcts.sptribs.ciccase.model.ApplicationType.JOINT_APPLICATION;
+import static uk.gov.hmcts.sptribs.ciccase.model.GeneralOrderJudgeOrLegalAdvisorType.ASSISTANT_JUSTICES_CLERK;
+import static uk.gov.hmcts.sptribs.ciccase.model.GeneralOrderJudgeOrLegalAdvisorType.PROPER_OFFICER_OF_THE_COURT;
+import static uk.gov.hmcts.sptribs.document.content.DocmosisTemplateConstants.GENERAL_ORDER_MADE_BY;
+import static uk.gov.hmcts.sptribs.testutil.TestConstants.TEST_CASE_ID;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.caseData;
+import static uk.gov.hmcts.sptribs.testutil.TestDataHelper.getGeneralOrder;
 
 
 @ExtendWith(SpringExtension.class)
